requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
openpyxl
celery[redis]
redis
python-dotenv
